fix(supabase): align edge function entrypoints for deploy
Objective:
- repair the remaining Supabase deploy failure on dev after migrations and per-function enumeration were already fixed

Changes:
- renamed tracked Supabase Edge Function entrypoints and shared helpers from .js to .ts so the CLI can find index.ts files
- updated shared imports and documented the index.ts deploy convention for Edge Functions
- excluded supabase/functions from the root Next.js TypeScript program so Deno-targeted sources do not break app typecheck

Validation:
- pnpm dlx node@22.22.1 /opt/homebrew/bin/pnpm lint
- pnpm dlx node@22.22.1 /opt/homebrew/bin/pnpm typecheck
- git diff --check
- find supabase/functions -mindepth 1 -maxdepth 2 -type f | sort

Notes:
- this is the third focused Supabase deploy repair after the migration idempotency fix and the per-function workflow deploy fix
- remote deploy confirmation still needs the follow-up PR merge and push-triggered dev workflow run
