+++ supabase/functions/_shared/command-runtime.ts
