+++ supabase/functions/project-crypto-route-enabled-set/index.ts
@@ -1,6 +1,6 @@
 import { validateProjectCryptoRouteEnabledSetInput } from "../../../lib/edge-functions/project-payment-operations-contract.ts"
 import { executeProjectCryptoRouteEnabledSetCommand } from "../../../lib/payments/project-payment-operations-command.ts"
-import { serveProjectPaymentOperation } from "../_shared/project-payment-operations.js"
+import { serveProjectPaymentOperation } from "../_shared/project-payment-operations.ts"
 
 const handleRequest = serveProjectPaymentOperation({
   validate: validateProjectCryptoRouteEnabledSetInput,
