+++ supabase/functions/project-crypto-route-update/index.ts
@@ -1,6 +1,6 @@
 import { validateProjectCryptoRouteUpdateInput } from "../../../lib/edge-functions/project-payment-operations-contract.ts"
 import { executeProjectCryptoRouteUpdateCommand } from "../../../lib/payments/project-payment-operations-command.ts"
-import { serveProjectPaymentOperation } from "../_shared/project-payment-operations.js"
+import { serveProjectPaymentOperation } from "../_shared/project-payment-operations.ts"
 
 const handleRequest = serveProjectPaymentOperation({
   validate: validateProjectCryptoRouteUpdateInput,
