+++ supabase/functions/project-onboarding-draft-upsert/index.ts
@@ -1,7 +1,7 @@
-import { validateProjectOnboardingPublishInput } from "../../../lib/edge-functions/project-onboarding-publish-contract.ts"
+import { validateProjectOnboardingDraftUpsertInput } from "../../../lib/edge-functions/project-onboarding-draft-upsert-contract.ts"
 import { edgeCommandFailure, edgeCommandSuccess } from "../../../lib/edge-functions/result.ts"
-import { executeProjectOnboardingPublishCommand } from "../../../lib/onboarding/project-onboarding-commands.ts"
-import { authenticateRequest, corsHeaders, json, parseJsonBody, serve } from "../_shared/command-runtime.js"
+import { executeProjectOnboardingDraftUpsertCommand } from "../../../lib/onboarding/project-onboarding-commands.ts"
+import { authenticateRequest, corsHeaders, json, parseJsonBody, serve } from "../_shared/command-runtime.ts"
 
 async function handleRequest(request) {
   if (request.method === "OPTIONS") {
@@ -13,7 +13,11 @@ async function handleRequest(request) {
   }
 
   const bodyResult = await parseJsonBody(request)
-  const validation = validateProjectOnboardingPublishInput(bodyResult.ok ? bodyResult.body : undefined)
+  if (!bodyResult.ok) {
+    return json(edgeCommandFailure("invalid_payload", bodyResult.error))
+  }
+
+  const validation = validateProjectOnboardingDraftUpsertInput(bodyResult.body)
   if (!validation.ok) {
     return json(validation)
   }
@@ -23,9 +27,10 @@ async function handleRequest(request) {
     return json(edgeCommandFailure(auth.code ?? "function_not_configured", auth.error))
   }
 
-  const result = await executeProjectOnboardingPublishCommand(auth.adminClient, {
+  const result = await executeProjectOnboardingDraftUpsertCommand(auth.adminClient, {
     actorUserId: auth.user.id,
-    rpcSupabase: auth.authClient,
+    currentScreen: validation.data.currentScreen,
+    payload: validation.data.payload,
   })
 
   if (!result.ok) {
