+++ supabase/functions/project-onboarding-publish/index.ts
@@ -1,7 +1,7 @@
-import { validateProjectOnboardingDraftUpsertInput } from "../../../lib/edge-functions/project-onboarding-draft-upsert-contract.ts"
+import { validateProjectOnboardingPublishInput } from "../../../lib/edge-functions/project-onboarding-publish-contract.ts"
 import { edgeCommandFailure, edgeCommandSuccess } from "../../../lib/edge-functions/result.ts"
-import { executeProjectOnboardingDraftUpsertCommand } from "../../../lib/onboarding/project-onboarding-commands.ts"
-import { authenticateRequest, corsHeaders, json, parseJsonBody, serve } from "../_shared/command-runtime.js"
+import { executeProjectOnboardingPublishCommand } from "../../../lib/onboarding/project-onboarding-commands.ts"
+import { authenticateRequest, corsHeaders, json, parseJsonBody, serve } from "../_shared/command-runtime.ts"
 
 async function handleRequest(request) {
   if (request.method === "OPTIONS") {
@@ -13,11 +13,7 @@ async function handleRequest(request) {
   }
 
   const bodyResult = await parseJsonBody(request)
-  if (!bodyResult.ok) {
-    return json(edgeCommandFailure("invalid_payload", bodyResult.error))
-  }
-
-  const validation = validateProjectOnboardingDraftUpsertInput(bodyResult.body)
+  const validation = validateProjectOnboardingPublishInput(bodyResult.ok ? bodyResult.body : undefined)
   if (!validation.ok) {
     return json(validation)
   }
@@ -27,10 +23,9 @@ async function handleRequest(request) {
     return json(edgeCommandFailure(auth.code ?? "function_not_configured", auth.error))
   }
 
-  const result = await executeProjectOnboardingDraftUpsertCommand(auth.adminClient, {
+  const result = await executeProjectOnboardingPublishCommand(auth.adminClient, {
     actorUserId: auth.user.id,
-    currentScreen: validation.data.currentScreen,
-    payload: validation.data.payload,
+    rpcSupabase: auth.authClient,
   })
 
   if (!result.ok) {
