+++ supabase/functions/user-onboarding-publish/index.ts
@@ -1,7 +1,7 @@
-import { validateUserOnboardingDraftUpsertInput } from "../../../lib/edge-functions/user-onboarding-draft-upsert-contract.ts"
+import { validateUserOnboardingPublishInput } from "../../../lib/edge-functions/user-onboarding-publish-contract.ts"
 import { edgeCommandFailure, edgeCommandSuccess } from "../../../lib/edge-functions/result.ts"
-import { executeUserOnboardingDraftUpsertCommand } from "../../../lib/onboarding/user-onboarding-commands.ts"
-import { authenticateRequest, corsHeaders, json, parseJsonBody, serve } from "../_shared/command-runtime.js"
+import { executeUserOnboardingPublishCommand } from "../../../lib/onboarding/user-onboarding-commands.ts"
+import { authenticateRequest, corsHeaders, json, parseJsonBody, serve } from "../_shared/command-runtime.ts"
 
 async function handleRequest(request) {
   if (request.method === "OPTIONS") {
@@ -13,11 +13,7 @@ async function handleRequest(request) {
   }
 
   const bodyResult = await parseJsonBody(request)
-  if (!bodyResult.ok) {
-    return json(edgeCommandFailure("invalid_payload", bodyResult.error))
-  }
-
-  const validation = validateUserOnboardingDraftUpsertInput(bodyResult.body)
+  const validation = validateUserOnboardingPublishInput(bodyResult.ok ? bodyResult.body : undefined)
   if (!validation.ok) {
     return json(validation)
   }
@@ -27,10 +23,9 @@ async function handleRequest(request) {
     return json(edgeCommandFailure(auth.code ?? "function_not_configured", auth.error))
   }
 
-  const result = await executeUserOnboardingDraftUpsertCommand(auth.adminClient, {
+  const result = await executeUserOnboardingPublishCommand(auth.adminClient, {
     actorUserId: auth.user.id,
-    currentScreen: validation.data.currentScreen,
-    payload: validation.data.payload,
+    actorEmail: auth.user.email ?? null,
   })
 
   if (!result.ok) {
